fix: separates exports of inMemEntityStore and inMemEntityService;removes public export of the store
