+++ packages/core/src/entity/index.ts
@@ -1,9 +1,9 @@
 export * from './EditorModeEntityStore';
 export * from './EntityStore';
 export * from './EntityStoreBase';
+
 export {
   inMemoryEntities,
-  inMemoryEntitiesStore,
   useInMemoryEntities,
   type InMemoryEntitiesService,
-} from './InMemoryEntities';
+} from './InMemoryEntitiesService';
